ADAS-34-refactor/컨트롤러/이성현: controller_ACC_LKAS_LCA구현rebase

+++ code/controller/ap/core/car_distance.h
@@ -54,7 +54,7 @@
 /*********************************************************************************************************************/
 
 void set_car_distance(int count);
-int get_car_distance(int count);
+float get_car_distance(int count);
 
 
 #endif /* AP_CORE_CAR_DISTANCE_H_ */
